+++ src/es_utils/reflection.rs
@@ -17,7 +17,7 @@ use mozjs::jsval::{JSVal, NullValue, ObjectValue, UndefinedValue};
 use mozjs::rust::{HandleObject, HandleValue};
 
 use crate::esvaluefacade::EsValueFacade;
-use std::borrow::BorrowMut;
+use std::borrow::{Borrow, BorrowMut};
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
 use std::ptr::replace;
@@ -27,7 +27,7 @@ use std::sync::Arc;
 ///
 
 pub struct Proxy {
-    class_name: String,
+    class_name: &'static str,
     constructor: Option<Box<dyn Fn(*mut JSContext, &CallArgs) -> Result<i32, String>>>,
     finalizer: Option<Box<dyn Fn(&i32) -> ()>>,
     properties: HashMap<
@@ -39,10 +39,16 @@ pub struct Proxy {
     >,
     methods: HashMap<&'static str, Box<dyn Fn(i32, &CallArgs) -> JSVal>>,
     events: HashSet<&'static str>,
+    event_listeners: RefCell<HashMap<i32, HashMap<&'static str, Vec<EsPersistentRooted>>>>,
+    static_properties:
+        HashMap<&'static str, (Box<dyn Fn() -> JSVal>, Box<dyn Fn(HandleValue) -> ()>)>,
+    static_methods: HashMap<&'static str, Box<dyn Fn(&CallArgs) -> JSVal>>,
+    static_events: HashSet<&'static str>,
+    static_event_listeners: RefCell<HashMap<&'static str, Vec<EsPersistentRooted>>>,
 }
 
 pub struct ProxyBuilder {
-    class_name: String,
+    class_name: &'static str,
     constructor: Option<Box<dyn Fn(*mut JSContext, &CallArgs) -> Result<i32, String>>>,
     finalizer: Option<Box<dyn Fn(&i32) -> ()>>,
     properties: HashMap<
@@ -54,18 +60,22 @@ pub struct ProxyBuilder {
     >,
     methods: HashMap<&'static str, Box<dyn Fn(i32, &CallArgs) -> JSVal>>,
     events: HashSet<&'static str>,
+    static_properties:
+        HashMap<&'static str, (Box<dyn Fn() -> JSVal>, Box<dyn Fn(HandleValue) -> ()>)>,
+    static_methods: HashMap<&'static str, Box<dyn Fn(&CallArgs) -> JSVal>>,
+    static_events: HashSet<&'static str>,
 }
 
 thread_local! {
     static PROXY_INSTANCE_IDS: RefCell<HashMap<usize, i32>> = RefCell::new(HashMap::new());
     static PROXY_INSTANCE_CLASSNAMES: RefCell<HashMap<i32, String>> = RefCell::new(HashMap::new());
-    static PROXIES: RefCell<HashMap<String, Arc<Proxy>>> = RefCell::new(HashMap::new());
+    static PROXIES: RefCell<HashMap<&'static str, Arc<Proxy>>> = RefCell::new(HashMap::new());
 }
 
 pub fn get_proxy(name: &str) -> Option<Arc<Proxy>> {
     // get proxy from PROXIES
-    PROXIES.with(|rc: &RefCell<HashMap<String, Arc<Proxy>>>| {
-        let map: &HashMap<String, Arc<Proxy>> = &*rc.borrow();
+    PROXIES.with(|rc: &RefCell<HashMap<&'static str, Arc<Proxy>>>| {
+        let map: &HashMap<&str, Arc<Proxy>> = &*rc.borrow();
         map.get(name).cloned()
     })
 }
@@ -73,12 +83,17 @@ pub fn get_proxy(name: &str) -> Option<Arc<Proxy>> {
 impl Proxy {
     fn new(cx: *mut JSContext, scope: HandleObject, builder: &mut ProxyBuilder) -> Arc<Self> {
         let mut ret = Proxy {
-            class_name: builder.class_name.to_string(),
+            class_name: builder.class_name,
             constructor: unsafe { replace(&mut builder.constructor, None) },
             finalizer: unsafe { replace(&mut builder.finalizer, None) },
             properties: HashMap::new(),
             methods: HashMap::new(),
             events: HashSet::new(),
+            event_listeners: RefCell::new(HashMap::new()),
+            static_properties: HashMap::new(),
+            static_methods: HashMap::new(),
+            static_events: HashSet::new(),
+            static_event_listeners: RefCell::new(HashMap::new()),
         };
 
         builder.properties.drain().all(|e| {
@@ -96,20 +111,46 @@ impl Proxy {
             true
         });
 
-        // todo, do we create an instance of proxy and set its constructor or do we define constructor and then add all static methods to that constructor obj? plan b is simpler but does that work on a native function?
+        builder.static_properties.drain().all(|e| {
+            ret.static_properties.insert(e.0, e.1);
+            true
+        });
 
-        crate::es_utils::functions::define_native_constructor(
-            cx,
-            scope,
-            ret.class_name.as_str(),
-            Some(construct),
-        );
+        builder.static_methods.drain().all(|e| {
+            ret.static_methods.insert(e.0, e.1);
+            true
+        });
+
+        builder.static_events.drain().all(|evt_type| {
+            ret.static_events.insert(evt_type);
+            true
+        });
+
+        // todo if no constructor: create an object instead of a function
+
+        let func: *mut mozjs::jsapi::JSFunction =
+            crate::es_utils::functions::define_native_constructor(
+                cx,
+                scope,
+                ret.class_name,
+                Some(proxy_construct),
+            );
+
+        ret.init_static_properties(cx, unsafe {
+            mozjs::rust::HandleObject::from_marked_location(&(func as *mut JSObject))
+        });
+        ret.init_static_methods(cx, unsafe {
+            mozjs::rust::HandleObject::from_marked_location(&(func as *mut JSObject))
+        });
+        ret.init_static_events(cx, unsafe {
+            mozjs::rust::HandleObject::from_marked_location(&(func as *mut JSObject))
+        });
 
         let ret_arc = Arc::new(ret);
 
-        PROXIES.with(|map_rc: &RefCell<HashMap<String, Arc<Proxy>>>| {
+        PROXIES.with(|map_rc: &RefCell<HashMap<&'static str, Arc<Proxy>>>| {
             let map = &mut *map_rc.borrow_mut();
-            map.insert(ret_arc.class_name.clone(), ret_arc.clone());
+            map.insert(ret_arc.class_name, ret_arc.clone());
         });
 
         ret_arc
@@ -120,12 +161,13 @@ impl Proxy {
     }
 
     pub fn dispatch_event(
+        &self,
         obj_id: i32,
         event_name: &str,
         cx: *mut JSContext,
         event_obj: mozjs::jsapi::HandleValue,
     ) {
-        dispatch_event_for_proxy(cx, obj_id, event_name, event_obj);
+        dispatch_event_for_proxy(cx, self, obj_id, event_name, event_obj);
     }
 
     pub fn dispatch_static_event(event_name: &str, event_obj: HandleObject) {
@@ -170,9 +212,9 @@ impl Proxy {
         panic!("NYI");
     }
 
-    /*fn init_properties(&self, cx: *mut JSContext, func: HandleObject) {
+    fn init_static_properties(&self, cx: *mut JSContext, func: HandleObject) {
         // this is actually how static_props should work, not instance props.. they should be resolved from the proxy_op
-        for prop_name in self.properties.keys() {
+        for prop_name in self.static_properties.keys() {
             // https://doc.servo.org/mozjs/jsapi/fn.JS_DefineProperty1.html
             // mozjs::jsapi::JS_DefineProperty1
             // todo move this to es_utils::object
@@ -181,26 +223,64 @@ impl Proxy {
             let ok = unsafe {
                 mozjs::jsapi::JS_DefineProperty1(
                     cx,
-                    func,
+                    func.into(),
                     n.as_ptr() as *const libc::c_char,
-                    JSPROP_PERMANENT + JSPROP_GETTER + JSPROP_SETTER + JSPROP_SHARED,
-                    Some(getter),
-                    Some(setter),
+                    Some(proxy_static_getter),
+                    Some(proxy_static_setter),
+                    (mozjs::jsapi::JSPROP_PERMANENT
+                        & mozjs::jsapi::JSPROP_GETTER
+                        & mozjs::jsapi::JSPROP_SETTER) as u32,
                 )
             };
         }
-    }*/
+    }
+
+    fn init_static_methods(&self, cx: *mut JSContext, func: HandleObject) {
+        trace!("init static methods for {}", self.class_name);
+        for method_name in self.static_methods.keys() {
+            trace!("init static method {} for {}", method_name, self.class_name);
+            crate::es_utils::functions::define_native_function(
+                cx,
+                func,
+                method_name,
+                Some(proxy_static_method),
+            );
+        }
+    }
+    fn init_static_events(&self, cx: *mut JSContext, func: HandleObject) {
+        crate::es_utils::functions::define_native_function(
+            cx,
+            func,
+            "addEventListener",
+            Some(proxy_static_add_event_listener),
+        );
+        crate::es_utils::functions::define_native_function(
+            cx,
+            func,
+            "removeEventListener",
+            Some(proxy_static_remove_event_listener),
+        );
+        crate::es_utils::functions::define_native_function(
+            cx,
+            func,
+            "dispatchEvent",
+            Some(proxy_static_dispatch_event),
+        );
+    }
 }
 
 impl ProxyBuilder {
-    pub fn new(class_name: &str) -> Self {
+    pub fn new(class_name: &'static str) -> Self {
         ProxyBuilder {
-            class_name: class_name.to_string(),
+            class_name: class_name,
             constructor: None,
             finalizer: None,
             properties: HashMap::new(),
             methods: HashMap::new(),
             events: HashSet::new(),
+            static_properties: HashMap::new(),
+            static_methods: HashMap::new(),
+            static_events: HashSet::new(),
         }
     }
 
@@ -230,11 +310,29 @@ impl ProxyBuilder {
         self
     }
 
+    pub fn static_property<G, S>(&mut self, name: &'static str, getter: G, setter: S) -> &mut Self
+    where
+        G: Fn() -> JSVal + 'static,
+        S: Fn(HandleValue) -> () + 'static,
+    {
+        self.static_properties
+            .insert(name, (Box::new(getter), Box::new(setter)));
+        self
+    }
+
     pub fn method<M>(&mut self, name: &'static str, method: M) -> &mut Self
     where
         M: Fn(i32, &CallArgs) -> JSVal + 'static,
     {
         self.methods.insert(name, Box::new(method));
+        self
+    }
+
+    pub fn static_method<M>(&mut self, name: &'static str, method: M) -> &mut Self
+    where
+        M: Fn(&CallArgs) -> JSVal + 'static,
+    {
+        self.static_methods.insert(name, Box::new(method));
         self
     }
 
@@ -244,6 +342,11 @@ impl ProxyBuilder {
 
     pub fn event(&mut self, evt_type: &'static str) -> &mut Self {
         self.events.insert(evt_type);
+        self
+    }
+
+    pub fn static_event(&mut self, evt_type: &'static str) -> &mut Self {
+        self.static_events.insert(evt_type);
         self
     }
 }
@@ -303,8 +406,76 @@ mod tests {
                         .build(cx, global);
                     let esvf = sm_rt
                         .eval(
-                            "let tp_obj = new TestClass1('bar'); tp_obj.abc = 1; console.log('tp_obj.abc = %s', tp_obj.abc); let i = tp_obj.foo; tp_obj.foo = 987; tp_obj.methodA(1, 2, 3); tp_obj.methodB(true); tp_obj.addEventListener('saved', (evt) => {console.log('tp_obj was saved');}); tp_obj.dispatchEvent('saved', {}); tp_obj = null; i;",
+                            "// create a new instance of your Proxy\n\
+                                      let tp_obj = new TestClass1('bar'); \n\
+                                      // you can set props that are not proxied \n\
+                                      tp_obj.abc = 1; console.log('tp_obj.abc = %s', tp_obj.abc); \n\
+                                      // test you getter and setter\n\
+                                      let i = tp_obj.foo; tp_obj.foo = 987; \n\
+                                      // test your method\n\
+                                      tp_obj.methodA(1, 2, 3);tp_obj.methodA(1, 2, 3);tp_obj.methodA(1, 2, 3); \n\
+                                      tp_obj.methodB(true); \n\
+                                      // add an event listener\n\
+                                      tp_obj.addEventListener('saved', (evt) => {console.log('tp_obj was saved');}); \n\
+                                      // dispatch an event from script\n\
+                                      tp_obj.dispatchEvent('saved', {}); \n\
+                                      // allow you object to be GCed\n\
+                                      tp_obj = null; i;",
                             "test_proxy.es",
+                        )
+                        .ok()
+                        .unwrap();
+                    assert_eq!(&123, esvf.get_i32());
+                });
+            });
+            inner.do_in_es_runtime_thread_sync(|sm_rt: &SmRuntime| {
+                sm_rt.cleanup();
+            });
+        });
+    }
+
+    #[test]
+    fn test_static_proxy() {
+        log::info!("test_static_proxy");
+        let rt = crate::esruntimewrapper::tests::TEST_RT.clone();
+
+        rt.do_with_inner(|inner| {
+            inner.do_in_es_runtime_thread_sync(|sm_rt: &SmRuntime| {
+                sm_rt.do_with_jsapi(|_rt, cx, global| {
+                    let _proxy_arc = ProxyBuilder::new("TestClass2")
+                        .static_property("foo", || {
+                            debug!("static_proxy.get foo");
+                            Int32Value(123)
+                        }, |_val| {
+                            debug!("static_proxy.set foo");
+                        })
+                        .static_property("bar", || {
+                            Int32Value(456)
+                        }, |_val| {})
+                        .static_method("methodA", |args| {
+                            trace!("static_proxy.methodA called with {} args", args.argc_);
+                            UndefinedValue()
+                        })
+                        .static_method("methodB", |args| {
+                            trace!("static_proxy.methodB called with {} args", args.argc_);
+                            UndefinedValue()
+                        })
+                        .event("saved")
+                        .build(cx, global);
+                    let esvf = sm_rt
+                        .eval(
+                            "// you can set props that a re not proxied \n\
+                                      TestClass2.abc = 1; console.log('TestClass2.abc = %s', TestClass2.abc); \n\
+                                      // test you getter and setter\n\
+                                      let i = TestClass2.foo; TestClass2.foo = 987; \n\
+                                      // test your method\n\
+                                      TestClass2.methodA(1, 2, 3);TestClass2.methodA(1, 2, 3);TestClass2.methodA(1, 2, 3); \n\
+                                      TestClass2.methodB(true); \n\
+                                      // add an event listener\n\
+                                      TestClass2.addEventListener('saved', (evt) => {console.log('TestClass2 was saved');}); \n\
+                                      // dispatch an event from script\n\
+                                      TestClass2.dispatchEvent('saved', {}); i;",
+                            "test_static_proxy.es",
                         )
                         .ok()
                         .unwrap();
@@ -318,20 +489,14 @@ mod tests {
     }
 }
 
-// todo, test resolve methods and such
-// todo should this wrap a native variant? with callargs and such
-// and thus should i impl a native variant first in es_utils?
-// yes.. yes i should...
-// should allways impl https://developer.mozilla.org/en-US/docs/Web/API/EventTarget
-
 static ES_PROXY_CLASS_CLASS_OPS: JSClassOps = JSClassOps {
     addProperty: None,
     delProperty: None,
     enumerate: None,
     newEnumerate: None,
-    resolve: Some(resolve),
+    resolve: Some(proxy_instance_resolve),
     mayResolve: None,
-    finalize: Some(finalize),
+    finalize: Some(proxy_instance_finalize),
     call: None,
     hasInstance: None,
     construct: None,
@@ -348,7 +513,7 @@ static ES_PROXY_CLASS: JSClass = JSClass {
 };
 
 /// resolvea property, this means if we know how to handle a prop we define that prop ob the instance obj
-unsafe extern "C" fn resolve(
+unsafe extern "C" fn proxy_instance_resolve(
     cx: *mut JSContext,
     obj: mozjs::jsapi::HandleObject,
     key: mozjs::jsapi::HandleId,
@@ -366,7 +531,7 @@ unsafe extern "C" fn resolve(
     if let Some(class_name) = class_name_res.ok() {
         PROXIES.with(|proxies_rc| {
             let proxies = &*proxies_rc.borrow();
-            if let Some(proxy) = proxies.get(&class_name) {
+            if let Some(proxy) = proxies.get(class_name.as_str()) {
                 trace!("check proxy {} for {}", class_name, prop_name);
 
                 if prop_name.as_str().eq("addEventListener") {
@@ -377,7 +542,7 @@ unsafe extern "C" fn resolve(
                         cx,
                         robj,
                         "addEventListener",
-                        Some(add_event_listener),
+                        Some(proxy_instance_add_event_listener),
                     );
 
                     *resolved = true;
@@ -390,7 +555,7 @@ unsafe extern "C" fn resolve(
                         cx,
                         robj,
                         "removeEventListener",
-                        Some(remove_event_listener),
+                        Some(proxy_instance_remove_event_listener),
                     );
 
                     *resolved = true;
@@ -403,7 +568,7 @@ unsafe extern "C" fn resolve(
                         cx,
                         robj,
                         "dispatchEvent",
-                        Some(dispatch_event),
+                        Some(proxy_instance_dispatch_event),
                     );
 
                     *resolved = true;
@@ -422,8 +587,8 @@ unsafe extern "C" fn resolve(
                         cx,
                         obj,
                         n.as_ptr() as *const libc::c_char,
-                        Some(getter),
-                        Some(setter),
+                        Some(proxy_instance_getter),
+                        Some(proxy_instance_setter),
                         (mozjs::jsapi::JSPROP_PERMANENT
                             & mozjs::jsapi::JSPROP_GETTER
                             & mozjs::jsapi::JSPROP_SETTER) as u32,
@@ -447,7 +612,7 @@ unsafe extern "C" fn resolve(
                         cx,
                         robj,
                         prop_name.as_str(),
-                        Some(method),
+                        Some(proxy_instance_method),
                     );
 
                     *resolved = true;
@@ -460,21 +625,21 @@ unsafe extern "C" fn resolve(
     true
 }
 
-unsafe extern "C" fn getter(cx: *mut JSContext, argc: u32, vp: *mut mozjs::jsapi::Value) -> bool {
+unsafe extern "C" fn proxy_instance_getter(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
     trace!("reflection::getter");
 
     let args = CallArgs::from_vp(vp, argc);
     let thisv: mozjs::jsapi::Value = *args.thisv();
 
     if thisv.is_object() {
-        let obj_handle = mozjs::rust::HandleObject::from_marked_location(&thisv.to_object());
-        let cn_res = crate::es_utils::objects::get_es_obj_prop_val_as_string(
-            cx,
-            obj_handle,
-            PROXY_PROP_CLASS_NAME,
-        );
-        if let Some(class_name) = cn_res.ok() {
-            trace!("reflection::getter get for cn:{}", class_name);
+        if let Some(proxy) = get_proxy_for(cx, thisv.to_object()) {
+            let obj_handle = mozjs::rust::HandleObject::from_marked_location(&thisv.to_object());
+
+            trace!("reflection::getter get for cn:{}", proxy.class_name);
 
             let callee: *mut JSObject = args.callee();
             let prop_name_res = crate::es_utils::objects::get_es_obj_prop_val_as_string(
@@ -484,7 +649,11 @@ unsafe extern "C" fn getter(cx: *mut JSContext, argc: u32, vp: *mut mozjs::jsapi
             );
             if let Some(prop_name) = prop_name_res.ok() {
                 // lovely the name here is "get [propname]"
-                trace!("reflection::getter get {} for cn:{}", prop_name, class_name);
+                trace!(
+                    "reflection::getter get {} for cn:{}",
+                    prop_name,
+                    proxy.class_name
+                );
 
                 // get obj id
                 let obj_id = crate::es_utils::objects::get_es_obj_prop_val_as_i32(
@@ -496,22 +665,61 @@ unsafe extern "C" fn getter(cx: *mut JSContext, argc: u32, vp: *mut mozjs::jsapi
                 trace!(
                     "reflection::getter get {} for cn:{} for obj_id {}",
                     prop_name,
-                    class_name,
+                    proxy.class_name,
                     obj_id
                 );
 
                 let p_name = &prop_name[4..];
 
-                PROXIES.with(|proxies_rc| {
-                    let proxies = &*proxies_rc.borrow();
-                    if let Some(proxy) = proxies.get(&class_name).cloned() {
-                        if let Some(prop) = proxy.properties.get(p_name) {
-                            let js_val = prop.0(obj_id);
-                            trace!("got val for getter");
-                            args.rval().set(js_val);
-                        }
-                    }
-                });
+                if let Some(prop) = proxy.properties.get(p_name) {
+                    let js_val = prop.0(obj_id);
+                    trace!("got val for getter");
+                    args.rval().set(js_val);
+                }
+            }
+        }
+    }
+
+    true
+}
+
+unsafe extern "C" fn proxy_static_getter(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
+    trace!("reflection::static_getter");
+
+    let args = CallArgs::from_vp(vp, argc);
+    let thisv: mozjs::jsapi::Value = *args.thisv();
+
+    if thisv.is_object() {
+        if let Some(proxy) = get_static_proxy_for(cx, thisv.to_object()) {
+            let obj_handle = mozjs::rust::HandleObject::from_marked_location(&thisv.to_object());
+
+            trace!("reflection::static_getter get for cn:{}", proxy.class_name);
+
+            let callee: *mut JSObject = args.callee();
+            let prop_name_res = crate::es_utils::objects::get_es_obj_prop_val_as_string(
+                cx,
+                HandleObject::from_marked_location(&callee),
+                "name",
+            );
+            if let Some(prop_name) = prop_name_res.ok() {
+                // lovely the name here is "get [propname]"
+                trace!(
+                    "reflection::static_getter get {} for cn:{}",
+                    prop_name,
+                    proxy.class_name
+                );
+
+                let p_name = &prop_name[4..];
+
+                if let Some(prop) = proxy.static_properties.get(p_name) {
+                    let js_val = prop.0();
+                    trace!("got val for static_getter");
+                    args.rval().set(js_val);
+                }
             }
         }
     }
@@ -534,14 +742,31 @@ fn get_proxy_for(cx: *mut JSContext, obj: *mut JSObject) -> Option<Arc<Proxy>> {
     if let Some(class_name) = cn_res.ok() {
         return PROXIES.with(|proxies_rc| {
             let proxies = &*proxies_rc.borrow();
-            proxies.get(&class_name).cloned()
+            proxies.get(class_name.as_str()).cloned()
         });
     }
 
     None
 }
 
-unsafe extern "C" fn setter(cx: *mut JSContext, argc: u32, vp: *mut mozjs::jsapi::Value) -> bool {
+fn get_static_proxy_for(cx: *mut JSContext, obj: *mut JSObject) -> Option<Arc<Proxy>> {
+    let obj_handle = unsafe { mozjs::rust::HandleObject::from_marked_location(&obj) };
+    let cn_res = crate::es_utils::objects::get_es_obj_prop_val_as_string(cx, obj_handle, "name");
+    if let Some(class_name) = cn_res.ok() {
+        return PROXIES.with(|proxies_rc| {
+            let proxies = &*proxies_rc.borrow();
+            proxies.get(class_name.as_str()).cloned()
+        });
+    }
+
+    None
+}
+
+unsafe extern "C" fn proxy_instance_setter(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
     trace!("reflection::setter");
 
     let args = CallArgs::from_vp(vp, argc);
@@ -586,12 +811,71 @@ unsafe extern "C" fn setter(cx: *mut JSContext, argc: u32, vp: *mut mozjs::jsapi
     true
 }
 
-thread_local! {
-    // todo key og top map should also contain proxy class name
-    pub static PROXY_EVENT_LISTENERS: RefCell<HashMap<i32, HashMap<&'static str, Vec<EsPersistentRooted>>>> = RefCell::new(HashMap::new());
+unsafe extern "C" fn proxy_static_setter(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
+    trace!("reflection::static_setter");
+
+    let args = CallArgs::from_vp(vp, argc);
+    let thisv: mozjs::jsapi::Value = *args.thisv();
+
+    if thisv.is_object() {
+        if let Some(proxy) = get_static_proxy_for(cx, thisv.to_object()) {
+            trace!("reflection::static_setter get for cn:{}", &proxy.class_name);
+
+            let callee: *mut JSObject = args.callee();
+            let prop_name_res = crate::es_utils::objects::get_es_obj_prop_val_as_string(
+                cx,
+                HandleObject::from_marked_location(&callee),
+                "name",
+            );
+            if let Some(prop_name) = prop_name_res.ok() {
+                // lovely the name here is "set [propname]"
+                trace!("reflection::static_setter set {}", prop_name);
+
+                // strip "set " from propname
+                let p_name = &prop_name[4..];
+
+                if let Some(prop) = proxy.static_properties.get(p_name) {
+                    let val = HandleValue::from_marked_location(&args.index(0).get());
+
+                    trace!("reflection::static_setter setting val");
+                    prop.1(val);
+                }
+            }
+        }
+    }
+
+    true
 }
 
-unsafe extern "C" fn add_event_listener(
+unsafe extern "C" fn proxy_static_add_event_listener(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
+    true
+}
+
+unsafe extern "C" fn proxy_static_remove_event_listener(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
+    true
+}
+
+unsafe extern "C" fn proxy_static_dispatch_event(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
+    true
+}
+
+unsafe extern "C" fn proxy_instance_add_event_listener(
     cx: *mut JSContext,
     argc: u32,
     vp: *mut mozjs::jsapi::Value,
@@ -619,20 +903,18 @@ unsafe extern "C" fn add_event_listener(
                 // we need this so we can get a &'static str
                 let type_str = proxy.events.get(type_str.as_str()).unwrap().clone();
 
-                PROXY_EVENT_LISTENERS.with(|pel_rc| {
-                    let pel = &mut *pel_rc.borrow_mut();
-                    if !pel.contains_key(&obj_id) {
-                        pel.insert(obj_id, HashMap::new());
-                    }
-                    let obj_map = pel.get_mut(&obj_id).unwrap();
+                let pel = &mut *proxy.event_listeners.borrow_mut();
+                if !pel.contains_key(&obj_id) {
+                    pel.insert(obj_id, HashMap::new());
+                }
+                let obj_map = pel.get_mut(&obj_id).unwrap();
 
-                    if !obj_map.contains_key(type_str) {
-                        obj_map.insert(type_str, vec![]);
-                    }
+                if !obj_map.contains_key(type_str) {
+                    obj_map.insert(type_str, vec![]);
+                }
 
-                    let listener_vec = obj_map.get_mut(type_str).unwrap();
-                    listener_vec.push(listener_epr);
-                });
+                let listener_vec = obj_map.get_mut(type_str).unwrap();
+                listener_vec.push(listener_epr);
             }
         }
     }
@@ -640,7 +922,7 @@ unsafe extern "C" fn add_event_listener(
     true
 }
 
-unsafe extern "C" fn remove_event_listener(
+unsafe extern "C" fn proxy_instance_remove_event_listener(
     cx: *mut JSContext,
     argc: u32,
     vp: *mut mozjs::jsapi::Value,
@@ -666,31 +948,30 @@ unsafe extern "C" fn remove_event_listener(
                 // we need this so we can get a &'static str
                 let type_str = proxy.events.get(type_str.as_str()).unwrap().clone();
 
-                PROXY_EVENT_LISTENERS.with(|pel_rc| {
-                    let pel = &mut *pel_rc.borrow_mut();
-                    if pel.contains_key(&obj_id) {
-                        let obj_map = pel.get_mut(&obj_id).unwrap();
+                let pel = &mut *proxy.event_listeners.borrow_mut();
 
-                        if obj_map.contains_key(type_str) {
-                            let listener_vec = obj_map.get_mut(type_str).unwrap();
-                            for x in 0..listener_vec.len() {
-                                let epr = listener_vec.get(x).unwrap();
-                                if epr.get() == listener_obj {
-                                    trace!("remove event listener for {}", type_str);
-                                    listener_vec.remove(x);
-                                    break;
-                                }
+                if pel.contains_key(&obj_id) {
+                    let obj_map = pel.get_mut(&obj_id).unwrap();
+
+                    if obj_map.contains_key(type_str) {
+                        let listener_vec = obj_map.get_mut(type_str).unwrap();
+                        for x in 0..listener_vec.len() {
+                            let epr = listener_vec.get(x).unwrap();
+                            if epr.get() == listener_obj {
+                                trace!("remove event listener for {}", type_str);
+                                listener_vec.remove(x);
+                                break;
                             }
                         }
                     }
-                });
+                }
             }
         }
     }
     true
 }
 
-unsafe extern "C" fn dispatch_event(
+unsafe extern "C" fn proxy_instance_dispatch_event(
     cx: *mut JSContext,
     argc: u32,
     vp: *mut mozjs::jsapi::Value,
@@ -714,7 +995,7 @@ unsafe extern "C" fn dispatch_event(
             if proxy.events.contains(&type_str.as_str()) {
                 let type_str = proxy.events.get(type_str.as_str()).unwrap().clone();
 
-                dispatch_event_for_proxy(cx, obj_id, type_str, evt_obj_handle_val);
+                dispatch_event_for_proxy(cx, proxy.borrow(), obj_id, type_str, evt_obj_handle_val);
             }
         }
     }
@@ -724,42 +1005,45 @@ unsafe extern "C" fn dispatch_event(
 // proxy can call this from Proxy::dispatch_event with esvf.to_es_val()
 fn dispatch_event_for_proxy(
     cx: *mut JSContext,
+    proxy: &Proxy,
     obj_id: i32,
     evt_type: &str,
     evt_obj: mozjs::jsapi::HandleValue,
 ) {
-    PROXY_EVENT_LISTENERS.with(|pel_rc| {
-        let pel = &*pel_rc.borrow();
-        if let Some(obj_map) = pel.get(&obj_id) {
-            if let Some(listener_vec) = obj_map.get(evt_type) {
-                rooted!(in (cx) let mut ret_val = UndefinedValue());
-                // todo this_obj should be the proxy obj..
-                rooted!(in (cx) let this_obj = NullValue().to_object_or_null());
-                // since evt_obj is already rooted here we don;t need the auto_root macro, we can just use call_method_value()
+    let pel = &*proxy.event_listeners.borrow();
+    if let Some(obj_map) = pel.get(&obj_id) {
+        if let Some(listener_vec) = obj_map.get(evt_type) {
+            rooted!(in (cx) let mut ret_val = UndefinedValue());
+            // todo this_obj should be the proxy obj..
+            rooted!(in (cx) let this_obj = NullValue().to_object_or_null());
+            // since evt_obj is already rooted here we don;t need the auto_root macro, we can just use call_method_value()
 
-                for listener_epr in listener_vec {
-                    let mut args_vec = vec![];
-                    args_vec.push(*evt_obj);
-                    let func_obj = listener_epr.get();
-                    // todo why do we only have a call_method by val and not by HandleObject?
-                    // the whole rooting func here could be avoided
-                    rooted!(in (cx) let function_val = ObjectValue(func_obj));
-                    crate::es_utils::functions::call_method_value(
-                        cx,
-                        this_obj.handle(),
-                        function_val.handle(),
-                        args_vec,
-                        ret_val.handle_mut(),
-                    )
-                    .ok()
-                    .unwrap();
-                }
+            for listener_epr in listener_vec {
+                let mut args_vec = vec![];
+                args_vec.push(*evt_obj);
+                let func_obj = listener_epr.get();
+                // todo why do we only have a call_method by val and not by HandleObject?
+                // the whole rooting func here could be avoided
+                rooted!(in (cx) let function_val = ObjectValue(func_obj));
+                crate::es_utils::functions::call_method_value(
+                    cx,
+                    this_obj.handle(),
+                    function_val.handle(),
+                    args_vec,
+                    ret_val.handle_mut(),
+                )
+                .ok()
+                .unwrap();
             }
         }
-    });
+    }
 }
 
-unsafe extern "C" fn method(cx: *mut JSContext, argc: u32, vp: *mut mozjs::jsapi::Value) -> bool {
+unsafe extern "C" fn proxy_instance_method(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
     trace!("reflection::method");
 
     let args = CallArgs::from_vp(vp, argc);
@@ -800,7 +1084,47 @@ unsafe extern "C" fn method(cx: *mut JSContext, argc: u32, vp: *mut mozjs::jsapi
     true
 }
 
-unsafe extern "C" fn finalize(_fop: *mut JSFreeOp, object: *mut JSObject) {
+unsafe extern "C" fn proxy_static_method(
+    cx: *mut JSContext,
+    argc: u32,
+    vp: *mut mozjs::jsapi::Value,
+) -> bool {
+    trace!("reflection::static_method");
+
+    let args = CallArgs::from_vp(vp, argc);
+    let thisv: mozjs::jsapi::Value = *args.thisv();
+
+    if thisv.is_object() {
+        if let Some(proxy) = get_static_proxy_for(cx, thisv.to_object()) {
+            let obj_handle = mozjs::rust::HandleObject::from_marked_location(&thisv.to_object());
+
+            trace!("reflection::static_method for cn:{}", &proxy.class_name);
+
+            let callee: *mut JSObject = args.callee();
+            let prop_name_res = crate::es_utils::objects::get_es_obj_prop_val_as_string(
+                cx,
+                HandleObject::from_marked_location(&callee),
+                "name",
+            );
+            if let Some(prop_name) = prop_name_res.ok() {
+                // lovely the name here is "get [propname]"
+                trace!("reflection::static_method {}", prop_name);
+
+                let p_name = prop_name.as_str();
+
+                if let Some(prop) = proxy.static_methods.get(p_name) {
+                    trace!("got method for static_method");
+                    let js_val = prop(&args); // todo pass Vec of HandleValue instead of callargs
+                    args.rval().set(js_val);
+                }
+            }
+        }
+    }
+
+    true
+}
+
+unsafe extern "C" fn proxy_instance_finalize(_fop: *mut JSFreeOp, object: *mut JSObject) {
     trace!("reflection::finalize");
 
     let ptr_usize = object as usize;
@@ -822,20 +1146,18 @@ unsafe extern "C" fn finalize(_fop: *mut JSFreeOp, object: *mut JSObject) {
             if let Some(finalizer) = &proxy.finalizer {
                 finalizer(&id);
             }
-        }
 
-        PROXY_EVENT_LISTENERS.with(|pel_rc| {
             // clear event listeners
-            let pel = &mut *pel_rc.borrow_mut();
+            let pel = &mut *proxy.event_listeners.borrow_mut();
             pel.remove(&id);
-        });
+        }
     }
 }
 
 const PROXY_PROP_CLASS_NAME: &str = "__proxy_class_name__";
 const PROXY_PROP_OBJ_ID: &str = "__proxy_obj_id__";
 
-unsafe extern "C" fn construct(
+unsafe extern "C" fn proxy_construct(
     cx: *mut JSContext,
     argc: u32,
     vp: *mut mozjs::jsapi::Value,